[hotfix] Extract PluginManager interface
