+++ flink-core/src/main/java/org/apache/flink/core/plugin/PluginManager.java
@@ -18,42 +18,13 @@
 
 package org.apache.flink.core.plugin;
 
-import org.apache.flink.annotation.Internal;
-
-import org.apache.flink.shaded.guava18.com.google.common.collect.Iterators;
-
-import javax.annotation.concurrent.ThreadSafe;
-
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collection;
 import java.util.Iterator;
 
 /**
- * Manager class and entry-point for the plugin mechanism in Flink.
+ * PluginManager is responsible for managing cluster plugins which are loaded using separate class loaders so that their dependencies
+ * don't interfere with Flink's dependencies.
  */
-@Internal
-@ThreadSafe
-public class PluginManager {
-
-	/** Parent-classloader to all classloader that are used for plugin loading. We expect that this is thread-safe. */
-	private final ClassLoader parentClassLoader;
-
-	/** A collection of descriptions of all plugins known to this plugin manager. */
-	private final Collection<PluginDescriptor> pluginDescriptors;
-
-	/** List of patterns for classes that should always be resolved from the parent ClassLoader. */
-	private final String[] alwaysParentFirstPatterns;
-
-	public PluginManager(Collection<PluginDescriptor> pluginDescriptors, String[] alwaysParentFirstPatterns) {
-		this(pluginDescriptors, PluginManager.class.getClassLoader(), alwaysParentFirstPatterns);
-	}
-
-	public PluginManager(Collection<PluginDescriptor> pluginDescriptors, ClassLoader parentClassLoader, String[] alwaysParentFirstPatterns) {
-		this.pluginDescriptors = pluginDescriptors;
-		this.parentClassLoader = parentClassLoader;
-		this.alwaysParentFirstPatterns = alwaysParentFirstPatterns;
-	}
+public interface PluginManager {
 
 	/**
 	 * Returns in iterator over all available implementations of the given service interface (SPI) in all the plugins
@@ -63,21 +34,5 @@ public class PluginManager {
 	 * @param <P> Type of the requested plugin service.
 	 * @return Iterator over all implementations of the given service that could be loaded from all known plugins.
 	 */
-	public <P> Iterator<P> load(Class<P> service) {
-		ArrayList<Iterator<P>> combinedIterators = new ArrayList<>(pluginDescriptors.size());
-		for (PluginDescriptor pluginDescriptor : pluginDescriptors) {
-			PluginLoader pluginLoader = PluginLoader.create(pluginDescriptor, parentClassLoader, alwaysParentFirstPatterns);
-			combinedIterators.add(pluginLoader.load(service));
-		}
-		return Iterators.concat(combinedIterators.iterator());
-	}
-
-	@Override
-	public String toString() {
-		return "PluginManager{" +
-			"parentClassLoader=" + parentClassLoader +
-			", pluginDescriptors=" + pluginDescriptors +
-			", alwaysParentFirstPatterns=" + Arrays.toString(alwaysParentFirstPatterns) +
-			'}';
-	}
+	<P> Iterator<P> load(Class<P> service);
 }
